pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-loop"
version = "0.1.0"
description = "AI-driven product iteration loop framework"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-tmp-files>=0.0.2",
    "pytest-cov>=4.0",
]
browser = [
    "playwright>=1.40",
]

[project.scripts]
ai-loop = "ai_loop.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=ai_loop --cov-report=term-missing --cov-fail-under=80"
